+++ src/asymmetric_crypto/curves/curve_25519/ed25519/certificate/mod.rs
@@ -2,7 +2,6 @@ mod encoding;
 
 use std::{str::FromStr, time::Duration};
 
-pub use encoding::ED25519_ALGORITHM_ID;
 use pkcs8::{
     der::{Decode, EncodePem},
     spki::SubjectPublicKeyInfoOwned,