+++ src/asymmetric_crypto/curves/curve_25519/ed25519/private_key.rs
@@ -1,5 +1,3 @@
-pub use ed25519_dalek::{SecretKey as EdSecretKey, VerifyingKey as EdPublicKey};
-
 use crate::Curve25519Secret;
 
 pub type Ed25519PrivateKey = Curve25519Secret;