+++ src/symmetric_crypto/key.rs
@@ -5,36 +5,40 @@ use std::ops::DerefMut;
 
 use zeroize::{Zeroize, ZeroizeOnDrop, Zeroizing};
 
+#[cfg(feature = "sha3")]
+use crate::kdf256;
 use crate::{
     reexport::rand_core::CryptoRngCore, CBytes, CryptoCoreError, FixedSizeCBytes,
-    RandomFixedSizeCBytes, SecretCBytes,
+    RandomFixedSizeCBytes, Secret, SecretCBytes,
 };
 
 /// A type that holds symmetric key of a fixed  size.
 ///
 /// It is internally built using an array of bytes of the given length.
 #[derive(Debug, Hash, PartialEq, Eq, Zeroize, ZeroizeOnDrop)]
-pub struct SymmetricKey<const LENGTH: usize>(pub(crate) [u8; LENGTH]);
+pub struct SymmetricKey<const LENGTH: usize>(Secret<LENGTH>);
 
 impl<const LENGTH: usize> CBytes for SymmetricKey<LENGTH> {}
 
 impl<const LENGTH: usize> FixedSizeCBytes<LENGTH> for SymmetricKey<LENGTH> {
     fn to_bytes(&self) -> [u8; LENGTH] {
-        self.0
+        let mut dest = [0; LENGTH];
+        self.0.to_unprotected_bytes(&mut dest);
+        dest
     }
 
-    fn try_from_bytes(bytes: [u8; LENGTH]) -> Result<Self, CryptoCoreError> {
-        Ok(Self(bytes))
+    fn try_from_bytes(mut bytes: [u8; LENGTH]) -> Result<Self, CryptoCoreError> {
+        Ok(Self(Secret::from_unprotected_bytes(&mut bytes)))
     }
 }
 
 impl<const LENGTH: usize> RandomFixedSizeCBytes<LENGTH> for SymmetricKey<LENGTH> {
+    #[inline(always)]
     fn new<R: CryptoRngCore>(rng: &mut R) -> Self {
-        let mut key = Self([0; LENGTH]);
-        rng.fill_bytes(&mut key.0);
-        key
+        Self(Secret::random(rng))
     }
 
+    #[inline(always)]
     fn as_bytes(&self) -> &[u8] {
         self.as_ref()
     }
@@ -59,13 +63,39 @@ impl<const LENGTH: usize> DerefMut for SymmetricKey<LENGTH> {
 
 impl<const LENGTH: usize> Default for SymmetricKey<LENGTH> {
     fn default() -> Self {
-        Self([0; LENGTH])
+        Self(Secret::new())
     }
 }
 
 impl<const LENGTH: usize> From<SymmetricKey<LENGTH>> for Zeroizing<Vec<u8>> {
     fn from(value: SymmetricKey<LENGTH>) -> Self {
         Zeroizing::new(value.0.to_vec())
+    }
+}
+
+#[cfg(feature = "sha3")]
+impl<const KEY_LENGTH: usize> SymmetricKey<KEY_LENGTH> {
+    /// Deterministically derive a new key from the given secret and additional information.
+    ///
+    /// # Error
+    ///
+    /// Fails to generate the key in case the secret evidently does not contain enough entropy. The
+    /// check performed is based on the respective key and secret lengths. The secret needs to be
+    /// generated from a source containing enough entropy (greater than its length) for this check
+    /// to be valid.
+    pub fn derive<const SECRET_LENGTH: usize>(
+        secret: &Secret<SECRET_LENGTH>,
+        info: &[u8],
+    ) -> Result<Self, CryptoCoreError> {
+        if SECRET_LENGTH < KEY_LENGTH {
+            return Err(CryptoCoreError::ConversionError(format!(
+                "insufficient entropy to derive {}-byte key from a {}-byte secret",
+                KEY_LENGTH, SECRET_LENGTH,
+            )));
+        }
+        let mut key = Self::default();
+        kdf256!(&mut key, &secret, info);
+        Ok(key)
     }
 }