+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## [9.4.0] - 2023-06-10
+
+### Features
+- add a `Secret` type to hold sensitive information
+- base the `SymmetricKey` type on the Secret type
+- add a derive constructor to the `SymmetricKey` that takes a `Secret` seed as
+  argument
+- add algebraic methods `one` and `zero` to the Curve25519 public key type.
+
 ## [9.3.0] - 2023-09-26
 
 ### Features